backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
reportlab==4.0.7
jinja2==3.1.2
pytest==7.4.3
httpx==0.25.2